refactor: generate stars as component
